# storage/swagger/http.json
{"openapi":"3.0.0","info":{"title":"MineAdmin","description":"MineAdmin 是一款基于 Hyperf 开发的开源管理系统，提供了用户管理、权限管理、系统设置、系统监控等功能。","termsOfService":"https:\/\/www.mineadmin.com","contact":{"name":"MineAdmin","url":"https:\/\/www.mineadmin.com\/about"},"license":{"name":"Apache2.0","url":"https:\/\/github.com\/mineadmin\/MineAdmin\/blob\/master\/LICENSE"},"version":"3.0.0"},"servers":[{"url":"http:\/\/127.0.0.1:9501","description":"本地服务"},{"url":"https:\/\/demo.mineadmin.com","description":"演示服务"}],"paths":{"\/admin\/passport\/login":{"post":{"tags":["admin:passport"],"summary":"系统登录","description":"登录.","operationId":"passportLogin","requestBody":{"content":{"application\/json":{"schema":{"properties":{"username":{"description":"用户名","type":"string","example":"admin","rules":"required|exists:user,username","attribute":null},"password":{"description":"密码","type":"string","example":"123456","rules":"required","attribute":null}},"type":"object"}}},"_content":{"properties":{"username":{"description":"用户名","type":"string","example":"admin","rules":"required|exists:user,username","attribute":null},"password":{"description":"密码","type":"string","example":"123456","rules":"required","attribute":null}},"type":"object"}},"responses":{"200":{"description":"登录成功","content":{"application\/json":{"schema":{"$ref":"#\/components\/schemas\/Result"},"example":"{\"code\":200,\"message\":\"成功\",\"data\":{\"token\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE3MjIwOTQwNTYsIm5iZiI6MTcyMjA5NDAiwiZXhwIjoxNzIyMDk0MzU2fQ.7EKiNHb_ZeLJ1NArDpmK6sdlP7NsDecsTKLSZn_3D7k\",\"expire_at\":300}}"}}}}}},"\/admin\/passport\/logout":{"post":{"tags":["admin:passport"],"summary":"退出","description":"退出.","operationId":"passportLogout","security":[{"bearerAuth":[]}]}},"getBingBackgroundImage":{"get":{"description":"获取每日的必应背景图","operationId":"getBingBackgroundImage","responses":{"200":{"description":"成功","content":{"application\/json":{"schema":[],"example":"{\n  \"images\": [\n    {\n      \"startdate\": \"20240726\",\n      \"fullstartdate\": \"202407261600\",\n      \"enddate\": \"20240727\",\n      \"url\": \"\/th?id=OHR.RhinelandVineyards_ZH-CN3332101688_1920x1080.jpg&rf=LaDigue_1920x1080.jpg&pid=hp\",\n      \"urlbase\": \"\/th?id=OHR.RhinelandVineyards_ZH-CN3332101688\",\n      \"copyright\": \"摩泽尔河谷的葡萄园，莱茵兰-法尔茨，德国 (© Jorg Greuel\/Getty Images)\",\n      \"copyrightlink\": \"https:\/\/www.bing.com\/search?q=%E6%B3%95%E5%B0%94%E8%8C%A8%E8%91%A1%E8%90%84%E9%85%92%E4%BA%A7%E5%8C%BA&form=hpcapt&mkt=zh-cn\",\n      \"title\": \"完美的葡萄酒\",\n      \"quiz\": \"\/search?q=Bing+homepage+quiz&filters=WQOskey:%22HPQuiz_20240726_RhinelandVineyards%22&FORM=HPQUIZ\",\n      \"wp\": true,\n      \"hsh\": \"4d0805d3edb368d9cebf56b7376cd938\",\n      \"drk\": 1,\n      \"top\": 1,\n      \"bot\": 1,\n      \"hs\": [\n        \n      ]\n    }\n  ],\n  \"tooltips\": {\n    \"loading\": \"正在加载...\",\n    \"previous\": \"上一个图像\",\n    \"next\": \"下一个图像\",\n    \"walle\": \"此图片不能下载用作壁纸。\",\n    \"walls\": \"下载今日美图。仅限用作桌面壁纸。\"\n  }\n}"}}}}}}},"components":{"schemas":{"User":{"title":"User","properties":{"id":{"title":"用户ID，主键","type":"integer","rules":null,"attribute":null},"username":{"title":"用户名","type":"string","rules":null,"attribute":null},"password":{"title":"密码","type":"string","rules":null,"attribute":null},"user_type":{"title":"用户类型：(100系统用户)","type":"string","rules":null,"attribute":null},"nickname":{"title":"用户昵称","type":"string","rules":null,"attribute":null},"phone":{"title":"手机","type":"string","rules":null,"attribute":null},"email":{"title":"用户邮箱","type":"string","rules":null,"attribute":null},"avatar":{"title":"用户头像","type":"string","rules":null,"attribute":null},"signed":{"title":"个人签名","type":"string","rules":null,"attribute":null},"dashboard":{"title":"后台首页类型","type":"string","rules":null,"attribute":null},"status":{"title":"状态 (1正常 2停用)","type":"integer","rules":null,"attribute":null},"login_ip":{"title":"最后登陆IP","type":"string","rules":null,"attribute":null},"login_time":{"title":"最后登陆时间","type":"mixed","rules":null,"attribute":null},"backend_setting":{"title":"后台设置数据","type":"string","rules":null,"attribute":null},"created_by":{"title":"创建者","type":"integer","rules":null,"attribute":null},"updated_by":{"title":"更新者","type":"integer","rules":null,"attribute":null},"created_at":{"title":"创建时间","type":"mixed","rules":null,"attribute":null},"updated_at":{"title":"更新时间","type":"mixed","rules":null,"attribute":null},"deleted_at":{"title":"删除时间","type":"mixed","rules":null,"attribute":null},"remark":{"title":"备注","type":"string","rules":null,"attribute":null}},"type":"object"},"ResultCode":{"title":"ResultCode","type":"integer","default":200,"enum":[200,500,401,403,404,405,406,422]},"Result":{"title":"Api Response","description":"Api Response","properties":{"code":{"$ref":"#\/components\/schemas\/ResultCode"},"message":{"title":"响应消息","type":"string","nullable":true,"rules":null,"attribute":null},"data":{"title":"响应数据","type":"mixed","nullable":true,"rules":null,"attribute":null}},"type":"object"},"PassportLoginVo":{"description":"登录成功返回","properties":{"token":{"description":"token","type":"string","example":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE3MjIwOTQwNTYsIm5iZiI6MTcyMjA5NDAiwiZXhwIjoxNzIyMDk0MzU2fQ.7EKiNHb_ZeLJ1NArDpmK6sdlP7NsDecsTKLSZn_3D7k","rules":null,"attribute":null},"expire_at":{"description":"过期时间,单位秒","type":"integer","example":300,"rules":null,"attribute":null}},"type":"object","example":"{\"token\":\"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpYXQiOjE3MjIwOTQwNTYsIm5iZiI6MTcyMjA5NDAiwiZXhwIjoxNzIyMDk0MzU2fQ.7EKiNHb_ZeLJ1NArDpmK6sdlP7NsDecsTKLSZn_3D7k\",\"expire_at\":300}"}},"securitySchemes":{"Bearer":{"type":"http","name":"Authorization","bearerFormat":"JWT","scheme":"bearer"},"ApiKey":{"type":"apiKey","name":"token","in":"header"}}},"externalDocs":{"description":"开发文档","url":"https:\/\/v3.doc.mineadmin.com"}}